LU-5478 osc: get rid of obd_* typedefs
We have a bunch of typedefs for common things that made no sense
and hid the actual type from plain view.
Replace them with proper uXX or sXX types.
Exception is in lustre_idl.h and lustre_ioctl.h where
they are replaced with __uXX and __sXX to be able to be included
in userspace

Also fix valid flags to u64.

patch 6 in series: modify osc/osp

Reviewed-on: http://review.whamcloud.com/13148
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/include/obd.h
@@ -1219,7 +1219,7 @@ static inline struct md_open_data *obd_mod_alloc(void)
 	}                                                         \
 })
 
-void obdo_from_inode(struct obdo *dst, struct inode *src, u32 valid);
+void obdo_from_inode(struct obdo *dst, struct inode *src, u64 valid);
 void obdo_set_parent_fid(struct obdo *dst, const struct lu_fid *parent);
 
 /* return 1 if client should be resend request */
