+++ lustre/include/obd_class.h
@@ -319,10 +319,10 @@ static inline struct lr_server_data *class_server_data(struct obd_device *obd)
 struct lu_attr;
 struct inode;
 
-void obdo_from_la(struct obdo *dst, const struct lu_attr *la, __u64 valid);
-void la_from_obdo(struct lu_attr *la, const struct obdo *dst, u32 valid);
+void obdo_from_la(struct obdo *dst, const struct lu_attr *la, u64 valid);
+void la_from_obdo(struct lu_attr *la, const struct obdo *dst, u64 valid);
 
-void obdo_cpy_md(struct obdo *dst, const struct obdo *src, u32 valid);
+void obdo_cpy_md(struct obdo *dst, const struct obdo *src, u64 valid);
 void obdo_to_ioobj(const struct obdo *oa, struct obd_ioobj *ioobj);
 
 #define OBT(dev)        (dev)->obd_type
