+++ lustre/obdclass/linux/linux-obdo.c
@@ -49,9 +49,9 @@
 #include <obd_class.h>
 
 /*FIXME: Just copy from obdo_from_inode*/
-void obdo_from_la(struct obdo *dst, const struct lu_attr *la, __u64 valid)
+void obdo_from_la(struct obdo *dst, const struct lu_attr *la, u64 valid)
 {
-	u32 newvalid = 0;
+	u64 newvalid = 0;
 
         if (valid & LA_ATIME) {
                 dst->o_atime = la->la_atime;
@@ -100,9 +100,9 @@ void obdo_from_la(struct obdo *dst, const struct lu_attr *la, __u64 valid)
 EXPORT_SYMBOL(obdo_from_la);
 
 /*FIXME: Just copy from obdo_from_inode*/
-void la_from_obdo(struct lu_attr *dst, const struct obdo *obdo, u32 valid)
+void la_from_obdo(struct lu_attr *dst, const struct obdo *obdo, u64 valid)
 {
-        __u64 newvalid = 0;
+	u64 newvalid = 0;
 
         valid &= obdo->o_valid;
